Successful meta-log build. Untested
Thank you mpkarpov for the starter code, that is helping plenty
