+++ MIDAS/src/data_logging_meta.h
@@ -1,5 +1,8 @@
 #pragma once
-#include "rocket_state.h"
+//#include "rocket_state.h"
+//#include "errors.h"
+
+/*
 
 #define META_LOGGING_MAX_SIZE 64
 
@@ -28,27 +31,12 @@ struct MetaLogging {
         size_t size;
     };
 
-    Queue<MetaLogEntry> _q;
+    bool get_queued(MetaLogEntry* out);
 
-    bool get_queued(MetaLogEntry* out) { return _q.receive(out); }
-
-    void log_event(MetaDataCode event_type, uint32_t timestamp) {
-        MetaLogEntry entry{event_type, 0, 0};
-        entry.size = sizeof(uint32_t);
-        memcpy(entry.data, &timestamp, entry.size);
-        _q.send(entry);
-    }
+    void log_event(MetaDataCode event_type, uint32_t timestamp);
 
     template <typename T>
-    void log_data(MetaDataCode data_type, const T& data) {
-
-        // double check...
-        static_assert(sizeof(T) <= META_LOGGING_MAX_SIZE, "Datatype for log_data too large");
-
-        MetaLogEntry entry{data_type, 0, 0};
-        entry.size = sizeof(T);
-        memcpy(entry.data, &data, entry.size);
-        _q.send(entry);
-    }
-
+    void log_data(MetaDataCode data_type, const T& data);
 };
+
+*/
